Update to mysql database
Added mysqlclient to Pipfile.
Updated project settings.py to use mysql engine and to read all of the
parameters from environment variables.
In pipenv, the easy way to automatically set up your environment is to
create a “.env” file in the project root directory that automatically will
set environment variables for this project when pipenv runs.